+++ client/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './App.css';
 import {BrowserRouter as Router} from 'react-router-dom'
-import Home from './pages';
+import Home from './pages/homePage.js';
 function App() {
   return (
     <Router>
