+++ client/src/components/Styles/footerElementsStyle.js
