+++ client/src/components/Styles/infoElementsStyle.js
