+++ client/src/components/Styles/mainElementsStyle.js
