+++ client/src/components/Styles/sidebarElementsStyle.js
