--- client/src/pages/index.js
@@ -1,29 +1,0 @@
-import React, {useState} from 'react'
-import Footer from '../components/Footer';
-import MainSection from '../components/MainSection';
-import InfoSection from '../components/InfoSection';
-import { homeObjOne, homeObjTwo ,homeObjThree } from '../components/InfoSection/Data';
-import Navbar from '../components/Navbar';
-import Services from '../components/Services';
-import Sidebar from '../components/Sidebar';
-const Home = () => {
-    const [isOpen, setIsOpen] = useState(false);
-    const toggle = () => {
-        setIsOpen(!isOpen);
-    }
-    return (
-        <>
-            <Sidebar isOpen={isOpen} toggle={toggle} />
-            <Navbar toggle={toggle} />
-            <MainSection />
-            <InfoSection {...homeObjOne} />
-            <InfoSection {...homeObjTwo} />
-            <Services />
-            <InfoSection {...homeObjThree} />
-            <Footer />
-
-        </>
-    )
-}
-
-export default Home;
